requirements.txt
boddle==0.2.8
bottle==0.12.13
certifi==2016.2.28
Paver==1.3.4
psycopg2==2.7.4
PyJWT==1.6.4
PyMySQL==0.8.1
python-dateutil==2.7.3
python-dotenv==0.8.2
six==1.11.0
SQLAlchemy==1.2.8
wincertstore==0.2
